+++ .gitignore
@@ -55,3 +55,5 @@ dependency-graph.png
 *.aux
 *.out
 *.synctex.gz
+/x/genutil/config/priv_validator_key.json
+/x/genutil/data/priv_validator_state.json
