+++ errors/stacktrace.go
@@ -73,7 +73,7 @@ func writeSimpleFrame(s io.Writer, f errors.Frame) {
 	if len(chunks) == 2 {
 		file = chunks[1]
 	}
-	fmt.Fprintf(s, " [%s:%d]", file, line)
+	_, _ = fmt.Fprintf(s, " [%s:%d]", file, line)
 }
 
 // Format works like pkg/errors, with additions.
@@ -86,16 +86,16 @@ func writeSimpleFrame(s io.Writer, f errors.Frame) {
 func (e *wrappedError) Format(s fmt.State, verb rune) {
 	// normal output here....
 	if verb != 'v' {
-		fmt.Fprint(s, e.Error())
+		_, _ = fmt.Fprint(s, e.Error())
 		return
 	}
 	// work with the stack trace... whole or part
 	stack := trimInternal(stackTrace(e))
 	if s.Flag('+') {
-		fmt.Fprintf(s, "%+v\n", stack)
-		fmt.Fprint(s, e.Error())
+		_, _ = fmt.Fprintf(s, "%+v\n", stack)
+		_, _ = fmt.Fprint(s, e.Error())
 	} else {
-		fmt.Fprint(s, e.Error())
+		_, _ = fmt.Fprint(s, e.Error())
 		writeSimpleFrame(s, stack[0])
 	}
 }
