+++ go.mod
@@ -12,6 +12,7 @@ require (
 	github.com/cosmos/btcutil v1.0.4
 	github.com/cosmos/cosmos-proto v0.0.0-20211123144845-528f5002c9f8
 	github.com/cosmos/cosmos-sdk/db v0.0.0
+	github.com/cosmos/cosmos-sdk/errors v0.0.0
 	github.com/cosmos/cosmos-sdk/x/group v0.0.0-00010101000000-000000000000
 	github.com/cosmos/go-bip39 v1.0.0
 	github.com/cosmos/iavl v0.17.3
@@ -156,3 +157,5 @@ replace github.com/gin-gonic/gin => github.com/gin-gonic/gin v1.7.0
 replace github.com/cosmos/cosmos-sdk/db => ./db
 
 replace github.com/cosmos/cosmos-sdk/x/group => ./x/group
+
+replace github.com/cosmos/cosmos-sdk/errors => ./errors
